first attempt at restructure
